Conflicts:
	gui/admin-gui/src/main/java/com/evolveum/midpoint/web/security/MidPointApplication.properties
	gui/admin-gui/src/main/java/com/evolveum/midpoint/web/util/ObjectTypeGuiDescriptor.java
	model/pom.xml

+++ repo/repo-sql-impl/src/main/java/com/evolveum/midpoint/repo/sql/data/common/id/RAExtDateId.java
@@ -25,7 +25,7 @@ import java.sql.Timestamp;
 public class RAExtDateId implements Serializable {
 
     private String ownerOid;
-    private Short ownerId;
+    private Integer ownerId;
     private Timestamp value;
     private String name;
 
@@ -37,11 +37,11 @@ public class RAExtDateId implements Serializable {
         this.ownerOid = ownerOid;
     }
 
-    public Short getOwnerId() {
+    public Integer getOwnerId() {
         return ownerId;
     }
 
-    public void setOwnerId(Short ownerId) {
+    public void setOwnerId(Integer ownerId) {
         this.ownerId = ownerId;
     }
